ChatGPT steals my job
ChatGPT wrote me a better script in 5 seconds then I could in a few hours. It made it as a text based health system, rather than a bar, but I think I prefer it that way. I moved all my old scripts to a new folder and applied the ChatGPT scripts to the characters in my scene.